create components FieldsetRellena
